requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
openai>=1.12.0
pydantic>=2.6.1
faiss-cpu>=1.12.0
sentence-transformers==2.3.1
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0
streamlit==1.31.1
plotly==5.18.0
